reduce diff between checkpoint v7 and checkpoint v6
